pyproject.toml
[project]
name = "ai-cv-customizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Usamahassan15", email = "hassanusama406@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "fastapi>=0.115.12",
    "fpdf>=1.7.2",
    "openai-agents>=0.0.14",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[project.scripts]
ai-cv-customizer = "ai_cv_customizer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
